video_llm/llava_api/llava-video-service/requirements.txt
aiohttp>=3.8.0
bentoml>=1.1.0
numpy>=1.24.0
torch>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9  # Using binary package instead of source
langsmith>=0.0.69  # For tracing
transformers>=4.36.0
git+https://github.com/LLaVA-VL/LLaVA-NeXT.git
decord>=0.6.0
Pillow>=10.0.0
